review - code organization - buildTableDescriptor + HardcodedSchema class extraction
